Rearranged and renamed subpackages to make more sense
