+++ operon/_cli/subcommands/install.py
@@ -2,10 +2,10 @@ import sys
 import os
 import shutil
 import argparse
-from operon.util.commands import BaseCommand
-import six.moves
-from operon.util import OperonState
 from datetime import datetime
+
+from operon._cli.subcommands import BaseSubcommand
+from operon._util.home import OperonState
 
 try:
     from pip import main as pip
@@ -18,19 +18,19 @@ EXIT_CMD_ERROR = 1
 EXIT_CMD_SYNTAX_ERROR = 2
 
 
-class Command(BaseCommand):
-    @staticmethod
-    def usage():
-        return 'operon install <pipeline-path> [-h]'
+def usage():
+    return 'operon install <pipeline-path> [-h]'
 
+
+class Subcommand(BaseSubcommand):
     def help_text(self):
         return 'Install a Operon formatted pipeline.'
 
-    def run(self, command_args):
-        parser = argparse.ArgumentParser(prog='operon install', usage=self.usage(), description=self.help_text())
+    def run(self, subcommand_args):
+        parser = argparse.ArgumentParser(prog='operon install', usage=usage(), description=self.help_text())
         parser.add_argument('pipeline-path', help='Full path to a Operon formatted pipeline to install.')
         parser.add_argument('-y', action='store_true', help='Install all dependency packages without asking.')
-        args = vars(parser.parse_args(command_args))
+        args = vars(parser.parse_args(subcommand_args))
         pipeline_path = args['pipeline-path']
         pipeline_name = os.path.splitext(os.path.basename(pipeline_path))[0]
         if pipeline_path.strip().lower() == 'help':
@@ -44,7 +44,7 @@ class Command(BaseCommand):
 
         # Check if pipeline already exists
         if os.path.isfile(os.path.join(self.home_pipelines, os.path.basename(pipeline_path))):
-            overwrite = six.moves.input('Pipeline {} is already installed, overwrite? [y/n] '.format(
+            overwrite = input('Pipeline {} is already installed, overwrite? [y/n] '.format(
                     os.path.basename(pipeline_path)
             ))
             if overwrite.lower() in {'no', 'n'}:
@@ -83,7 +83,7 @@ class Command(BaseCommand):
             pipeline_class._print_dependencies()
 
             if not args['y']:
-                install_depencencies = six.moves.input('\nProceed with dependency installation? [y/n] ')
+                install_depencencies = input('\nProceed with dependency installation? [y/n] ')
                 if install_depencencies.lower() in {'no', 'n', 'nope', 'nada', 'nah'}:
                     sys.exit(EXIT_CMD_SUCCESS)
             for package in pipeline_dependencies:
