+++ operon/_cli/subcommands/run.py
@@ -1,7 +1,8 @@
-import argparse
 import os
 import sys
-from operon.util.commands import BaseCommand
+import argparse
+
+from operon._cli.subcommands import BaseSubcommand
 
 ARGV_FIRST_ARGUMENT = 0
 ARGV_PIPELINE_NAME = 0
@@ -9,23 +10,23 @@ EXIT_CMD_SUCCESS = 0
 EXIT_CMD_SYNTAX_ERROR = 2
 
 
-class Command(BaseCommand):
-    @staticmethod
-    def usage():
-        return 'operon run <pipeline-name> [-h] [pipeline-options]'
+def usage():
+    return 'operon run <pipeline-name> [-h] [pipeline-options]'
 
+
+class Subcommand(BaseSubcommand):
     def help_text(self):
         return 'Run a pipeline with pipeline-specific parameters.'
 
-    def run(self, command_args):
+    def run(self, subcommand_args):
         # Get pipeline name or output help
-        parser = argparse.ArgumentParser(prog='operon run', usage=self.usage(), description=self.help_text())
-        if not command_args or command_args[ARGV_FIRST_ARGUMENT].lower() in ['-h', '--help', 'help']:
+        parser = argparse.ArgumentParser(prog='operon run', usage=usage(), description=self.help_text())
+        if not subcommand_args or subcommand_args[ARGV_FIRST_ARGUMENT].lower() in ['-h', '--help', 'help']:
             parser.print_help()
             sys.exit(EXIT_CMD_SUCCESS)
 
         # Get the pipeline class based on the name
-        pipeline_name = command_args[ARGV_PIPELINE_NAME]
+        pipeline_name = subcommand_args[ARGV_PIPELINE_NAME]
         pipeline_class = self.get_pipeline_class(pipeline_name)
 
         if pipeline_class is not None:
@@ -40,7 +41,7 @@ class Command(BaseCommand):
 
             # Get custom arguments from the Pipeline
             pipeline_class.add_pipeline_args(pipeline_args_parser)
-            pipeline_class.pipeline_args = vars(pipeline_args_parser.parse_args(command_args[1:]))
+            pipeline_class.pipeline_args = vars(pipeline_args_parser.parse_args(subcommand_args[1:]))
 
             # Parse pipeline config and run pipeline
             pipeline_class._parse_config()
